+++ varken/__init__.py
@@ -1,1 +1,2 @@
-VERSION = 1.1
+VERSION = 1.4
+BRANCH = 'master'