+++ varken/helpers.py
@@ -1,70 +1,82 @@
-import os
-import time
-import tarfile
-import hashlib
-import urllib3
-import geoip2.database
-import logging
-
-from json.decoder import JSONDecodeError
-from os.path import abspath, join
-from requests.exceptions import InvalidSchema, SSLError
+from hashlib import md5
+from datetime import date
+from logging import getLogger
+from calendar import monthcalendar
+from geoip2.database import Reader
+from tarfile import open as taropen
+from urllib3 import disable_warnings
+from os import stat, remove, makedirs
 from urllib.request import urlretrieve
+from json.decoder import JSONDecodeError
+from os.path import abspath, join, basename, isdir
+from urllib3.exceptions import InsecureRequestWarning
+from requests.exceptions import InvalidSchema, SSLError, ConnectionError
 
-logger = logging.getLogger('varken')
-
-
-def geoip_download(data_folder):
-    datafolder = data_folder
-
-    tar_dbfile = abspath(join(datafolder, 'GeoLite2-City.tar.gz'))
-
-    url = 'http://geolite.maxmind.com/download/geoip/database/GeoLite2-City.tar.gz'
-    logger.info('Downloading GeoLite2 from %s', url)
-    urlretrieve(url, tar_dbfile)
-
-    tar = tarfile.open(tar_dbfile, 'r:gz')
-    logging.debug('Opening GeoLite2 tar file : %s', tar_dbfile)
-
-    for files in tar.getmembers():
-        if 'GeoLite2-City.mmdb' in files.name:
-            logging.debug('"GeoLite2-City.mmdb" FOUND in tar file')
-            files.name = os.path.basename(files.name)
-
-            tar.extract(files, datafolder)
-            logging.debug('%s has been extracted to %s', files, datafolder)
-        
-    os.remove(tar_dbfile)
+logger = getLogger()
 
 
-def geo_lookup(ipaddress, data_folder):
-    datafolder = data_folder
-    logging.debug('Reading GeoLite2 from %s', datafolder)
+class GeoIPHandler(object):
+    def __init__(self, data_folder):
+        self.data_folder = data_folder
+        self.dbfile = abspath(join(self.data_folder, 'GeoLite2-City.mmdb'))
+        self.logger = getLogger()
+        self.update()
 
-    dbfile = abspath(join(datafolder, 'GeoLite2-City.mmdb'))
-    now = time.time()
+        self.logger.info('Opening persistent connection to GeoLite2 DB...')
+        self.reader = Reader(self.dbfile)
 
-    try:
-        dbinfo = os.stat(dbfile)
-        db_age = now - dbinfo.st_ctime
-        if db_age > (35 * 86400):
-            logging.info('GeoLite2 DB is older than 35 days. Attempting to re-download...')
+    def lookup(self, ipaddress):
+        ip = ipaddress
+        self.logger.debug('Getting lat/long for Tautulli stream')
+        return self.reader.city(ip)
 
-            os.remove(dbfile)
+    def update(self):
+        today = date.today()
+        dbdate = None
+        try:
+            dbdate = date.fromtimestamp(stat(self.dbfile).st_ctime)
+        except FileNotFoundError:
+            self.logger.error("Could not find GeoLite2 DB as: %s", self.dbfile)
+            self.download()
+        first_wednesday_day = [week[2:3][0] for week in monthcalendar(today.year, today.month) if week[2:3][0] != 0][0]
+        first_wednesday_date = date(today.year, today.month, first_wednesday_day)
 
-            geoip_download(datafolder)
-    except FileNotFoundError:
-        logging.error('GeoLite2 DB not found. Attempting to download...')
-        geoip_download(datafolder)
+        if dbdate < first_wednesday_date < today:
+            self.logger.info("Newer GeoLite2 DB available, Updating...")
+            remove(self.dbfile)
+            self.download()
+        else:
+            td = first_wednesday_date - today
+            if td.days < 0:
+                self.logger.debug('Geolite2 DB is only %s days old. Keeping current copy', abs(td.days))
+            else:
+                self.logger.debug('Geolite2 DB will update in %s days', abs(td.days))
 
-    reader = geoip2.database.Reader(dbfile)
 
-    return reader.city(ipaddress)
+    def download(self):
+        tar_dbfile = abspath(join(self.data_folder, 'GeoLite2-City.tar.gz'))
+        url = 'http://geolite.maxmind.com/download/geoip/database/GeoLite2-City.tar.gz'
+
+        self.logger.info('Downloading GeoLite2 from %s', url)
+        urlretrieve(url, tar_dbfile)
+
+        self.logger.debug('Opening GeoLite2 tar file : %s', tar_dbfile)
+
+        tar = taropen(tar_dbfile, 'r:gz')
+
+        for files in tar.getmembers():
+            if 'GeoLite2-City.mmdb' in files.name:
+                self.logger.debug('"GeoLite2-City.mmdb" FOUND in tar file')
+                files.name = basename(files.name)
+                tar.extract(files, self.data_folder)
+                self.logger.debug('%s has been extracted to %s', files, self.data_folder)
+        tar.close()
+        remove(tar_dbfile)
 
 
 def hashit(string):
     encoded = string.encode()
-    hashed = hashlib.md5(encoded).hexdigest()
+    hashed = md5(encoded).hexdigest()
 
     return hashed
 
@@ -75,37 +87,59 @@ def connection_handler(session, request, verify):
     v = verify
     return_json = False
 
-    urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
+    disable_warnings(InsecureRequestWarning)
 
     try:
         get = s.send(r, verify=v)
         if get.status_code == 401:
-            logger.info('Your api key is incorrect for {}'.format(r.url))
+            logger.info('Your api key is incorrect for %s', r.url)
         elif get.status_code == 404:
-            logger.info('This url doesnt even resolve: {}'.format(r.url))
+            logger.info('This url doesnt even resolve: %s', r.url)
         elif get.status_code == 200:
             try:
                 return_json = get.json()
             except JSONDecodeError:
-                logger.error('No JSON response... BORKED! Let us know in discord')
+                logger.error('No JSON response. Response is: %s', get.text)
         # 204 No Content is for ASA only
         elif get.status_code == 204:
             if get.headers['X-Auth-Token']:
                 return get.headers['X-Auth-Token']
 
     except InvalidSchema:
-        logger.error('You added http(s):// in the config file. Don\'t do that.')
+        logger.error("You added http(s):// in the config file. Don't do that.")
 
     except SSLError as e:
         logger.error('Either your host is unreachable or you have an SSL issue. : %s', e)
+
+    except ConnectionError as e:
+        logger.error('Cannot resolve the url/ip/port. Check connectivity. Error: %s', e)
 
     return return_json
 
 
 def mkdir_p(path):
-    """http://stackoverflow.com/a/600612/190597 (tzot)"""
+    templogger = getLogger('temp')
     try:
-        logger.info('Creating folder %s ', path)
-        os.makedirs(path, exist_ok=True)
+        if not isdir(path):
+            templogger.info('Creating folder %s ', path)
+            makedirs(path, exist_ok=True)
     except Exception as e:
-        logger.error('Could not create folder %s : %s ', path, e)
+        templogger.error('Could not create folder %s : %s ', path, e)
+
+
+def clean_sid_check(server_id_list, server_type=None):
+    t = server_type
+    sid_list = server_id_list
+    cleaned_list = sid_list.replace(' ', '').split(',')
+    valid_sids = []
+    for sid in cleaned_list:
+        try:
+            valid_sids.append(int(sid))
+        except ValueError:
+            logger.error("%s is not a valid server id number", sid)
+    if valid_sids:
+        logger.info('%s : %s', t.upper(), valid_sids)
+        return valid_sids
+    else:
+        logger.error('No valid %s', t.upper())
+        return False